DRY up nodejs test drivers
